Merge org.eclipse.swtchart.extensions.eclipse into org.eclipse.swtchart.extenssions
Having theming for elements being in separate bundle simply complicates
things for the sake of not drawing in E4 css bundles in. But the
extensions bundle already drives in org.eclipse.ui (and thus big(bigger)
part of Eclipse in so adding the css bundles makes it easier for
adapters (to get theming by default) and even simplifies releng.